refactor(#467): move HDR pipeline sources into src/HDR/
Group HDREnvironmentManager and HdrEquirectLoader (plus tests) under
src/HDR/ with PS1-style HDR/… includes for a clearer epic layout.
