+++ src/HDR/HdrEquirectLoader.h
